move ConfigFlags to pkg/kubectl/genericclioptions
